Attributes which have less-than or greater-than symbols are now properly output by the XML. No commas between attributes.
More information is printed out for enumerated types. In particular: if there is a typedef on an anonymous enum, the typedef will be output. If there is a tag, only the tag will be output. This gives us just enough information to find which enum an argument or parameter enum constant belongs to.