Created a matcher for struct declarations in which they must not be from an included file, and must be a definition of a struct (rather than forward-declaration)
Modified the enum matcher so that it only matches against enums which are definitions

Created custom AST_MATCHER functions to assist with matching against structs.

Created XML element factories for structs and field.

Wrote the handler for struct matchers.

Modified the makefile for the test target so that the output is tabulated.

Added a couple structs to the test header.